+++ CHANGES.rst
@@ -5,6 +5,9 @@
   does not have the target's proper motion / coordinate; or it is from
   `tpf.cutout()` call. [#1088]
 
+- Fixed a bug in ``TargetPixelFile.estimate_centroids`` which caused the column
+  and row coordinates reported to be off by 0.5.
+
 
 2.0.10 (2021-06-04)
 ===================